# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "App_meteorite Core Server",
    "description": "Back-end expert d'identification avec gestion flexible des flux multimédias",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/scan/exterior": {
      "post": {
        "summary": "Scan Exterior",
        "operationId": "scan_exterior_api_v1_scan_exterior_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_scan_exterior_api_v1_scan_exterior_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanDecisionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requete invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Cle API invalide ou manquante",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Quota epuise",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit metier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Fichier trop volumineux",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Format de fichier non supporte",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Erreur de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service indisponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scan/{scan_id}/interior": {
      "patch": {
        "summary": "Scan Interior Update",
        "operationId": "scan_interior_update_api_v1_scan__scan_id__interior_patch",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_scan_interior_update_api_v1_scan__scan_id__interior_patch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanDecisionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requete invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Cle API invalide ou manquante",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Quota epuise",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit metier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Fichier trop volumineux",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Format de fichier non supporte",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Erreur de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service indisponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/marketplace/publish/{scan_id}": {
      "post": {
        "summary": "Publish To Marketplace",
        "description": "Route de mise en vente sur le Marketplace ou validation finale.\nSécurise la vie privée via floutage des coordonnées.",
        "operationId": "publish_to_marketplace_api_v1_marketplace_publish__scan_id__post",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceListingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requete invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Cle API invalide ou manquante",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Quota epuise",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit metier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Fichier trop volumineux",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Format de fichier non supporte",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Erreur de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service indisponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "properties": {
          "status_code": {
            "type": "string",
            "title": "Status Code",
            "default": "DIAGNOSTIC_FAILED"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse"
      },
      "Body_scan_exterior_api_v1_scan_exterior_post": {
        "properties": {
          "client_uuid": {
            "type": "string",
            "title": "Client Uuid"
          },
          "files_exterior": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files Exterior"
          },
          "file_interior": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Interior"
          },
          "weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight"
          },
          "magnetic": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Magnetic"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "client_uuid",
          "files_exterior",
          "user_id"
        ],
        "title": "Body_scan_exterior_api_v1_scan_exterior_post"
      },
      "Body_scan_interior_update_api_v1_scan__scan_id__interior_patch": {
        "properties": {
          "file_interior": {
            "type": "string",
            "format": "binary",
            "title": "File Interior"
          }
        },
        "type": "object",
        "required": [
          "file_interior"
        ],
        "title": "Body_scan_interior_update_api_v1_scan__scan_id__interior_patch"
      },
...
      "ScanActions": {
        "properties": {
          "add_to_collection": {
            "type": "boolean",
            "title": "Add To Collection"
          },
          "enable_marketplace_button": {
            "type": "boolean",
            "title": "Enable Marketplace Button"
          },
          "invite_interior_cut": {
            "type": "boolean",
            "title": "Invite Interior Cut"
          }
        },
        "type": "object",
        "required": [
          "add_to_collection",
          "enable_marketplace_button",
          "invite_interior_cut"
        ],
        "title": "ScanActions"
      },
      "ScanDecisionResponse": {
        "properties": {
          "status_code": {
            "type": "string",
            "title": "Status Code"
          },
          "is_meteorite": {
            "type": "boolean",
            "title": "Is Meteorite"
          },
          "meteorite_probability": {
            "type": "number",
            "title": "Meteorite Probability"
          },
          "dominant_class": {
            "type": "string",
            "title": "Dominant Class"
          },
          "class_confidence": {
            "type": "number",
            "title": "Class Confidence"
          },
          "actions": {
            "$ref": "#/components/schemas/ScanActions"
          },
          "trigger_radar_admin": {
            "type": "boolean",
            "title": "Trigger Radar Admin"
          },
          "metadata_applied": {
            "$ref": "#/components/schemas/ScanMetadataApplied"
          },
          "scan_id": {
            "type": "string",
            "title": "Scan Id"
          },
          "is_sync_retry": {
            "type": "boolean",
            "title": "Is Sync Retry",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "status_code",
          "is_meteorite",
          "meteorite_probability",
          "dominant_class",
          "class_confidence",
          "actions",
          "trigger_radar_admin",
          "metadata_applied",
          "scan_id"
        ],
        "title": "ScanDecisionResponse"
      },
      "ScanMetadataApplied": {
        "properties": {
          "weight_provided": {
            "type": "boolean",
            "title": "Weight Provided"
          },
          "magnetic_status": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Magnetic Status"
          },
          "has_coordinates": {
            "type": "boolean",
            "title": "Has Coordinates"
          }
        },
        "type": "object",
        "required": [
          "weight_provided",
          "has_coordinates"
        ],
        "title": "ScanMetadataApplied"
      }
    }
  }
}